fix bug

+++ interface/api/rest/consoleController.go
@@ -26,8 +26,8 @@ func NewConsoleController(service interfaces.ConsoleService) *ConsoleController 
 func (gc *ConsoleController) SetupRoutes(router *gin.Engine) {
 	consoleGroup := router.Group("/consoles")
 	{
-		consoleGroup.Use(middleware.AuthMiddleware(&auth.AuthService{}))
 		consoleGroup.GET("", gc.GetAllConsolesHandler)
+		consoleGroup.Use(middleware.AuthMiddleware(&auth.AuthService{}))
 		consoleGroup.POST("/addConsole", gc.AddConsoleHandler)
 		consoleGroup.PUT("/updateConsole", gc.UpdateConsoleHandler)
 		consoleGroup.DELETE("/deleteConsole/:id", gc.DeleteConsole)
@@ -46,6 +46,16 @@ func (controller *ConsoleController) GetAllConsolesHandler(c *gin.Context) {
 }
 
 func (controller *ConsoleController) AddConsoleHandler(c *gin.Context) {
+	role, exists := c.Get("role")
+	if !exists {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "User not found in context"})
+		return
+	}
+	if role != "kro-admin" {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "You don't have permission"})
+		return
+	}
+
 	var newConsole restmodel.Console
 	if err := c.ShouldBindJSON(&newConsole); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -62,6 +72,15 @@ func (controller *ConsoleController) AddConsoleHandler(c *gin.Context) {
 }
 
 func (controller *ConsoleController) UpdateConsoleHandler(c *gin.Context) {
+	role, exists := c.Get("role")
+	if !exists {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "User not found in context"})
+		return
+	}
+	if role != "kro-admin" {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "You don't have permission"})
+		return
+	}
 	var updatedConsole entity.Console
 	if err := c.ShouldBindJSON(&updatedConsole); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -79,7 +98,15 @@ func (controller *ConsoleController) UpdateConsoleHandler(c *gin.Context) {
 
 func (controller *ConsoleController) DeleteConsole(c *gin.Context) {
 	id := c.Param("id")
-
+	role, exists := c.Get("role")
+	if !exists {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "User not found in context"})
+		return
+	}
+	if role != "kro-admin" {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "You don't have permission"})
+		return
+	}
 	err := controller.service.DeleteConsole(id)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete console"})
